added exception handling

+++ WardrobeOrganizer/Controllers/AccessoriesController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using System.Data;
 using System.Drawing;
@@ -9,6 +10,7 @@ using WardrobeOrganizer.Core.Models.Outerwear;
 using WardrobeOrganizer.Core.Services;
 using WardrobeOrganizer.Extensions;
 using WardrobeOrganizer.Infrastructure.Data;
+using WardrobeOrganizer.Infrastructure.Data.Enums;
 
 namespace WardrobeOrganizer.Controllers
 {
@@ -17,28 +19,77 @@ namespace WardrobeOrganizer.Controllers
         private readonly IAccessoriesService accessoriesService;
         private readonly IFamilyService familyService;
         private readonly IMemberService memberService;
+        private readonly IStorageService storageService;
+        private readonly IHouseService houseService;
+        private readonly UserManager<User> userManager;
 
 
         public AccessoriesController(IAccessoriesService _accessoriesService, 
             IFamilyService _familyService, 
-            IMemberService _memberService)
+            IMemberService _memberService,
+            IStorageService _storageService,
+            UserManager<User> _userManager)
         {
             this.accessoriesService = _accessoriesService;
             this.familyService = _familyService;
             this.memberService = _memberService;
+            this.storageService = _storageService;
         }
 
         public async Task<IActionResult> All(int storageId)
-        {         
-            var model = await accessoriesService.AllAccessories(storageId);
-            return View(model);
+        {
+            if (await storageService.ExistsById(storageId) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid storage";
+                return RedirectToAction("Error", "Home");
+            }
+            var storage = await storageService.GetStorageById(storageId);
+            int familiId = await familyService.GetFamilyId(User.Id());
+
+            if (storage.House.FamilyId != familiId)
+            {
+                TempData[MessageConstant.WarningMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
+            try
+            {
+                var model = await accessoriesService.AllAccessories(storageId);
+                return View(model); 
+            }
+            catch (Exception)
+            {
+
+                return RedirectToAction("Error", "Home");
+            }
+           
         }
 
         [HttpGet]
         [Authorize(Roles = RoleConstants.User)]
         public async Task<IActionResult> Add(int storageId, string category)
         {
-            var familyId = await familyService.GetFamilyId(User.Id());
+            if (await storageService.ExistsById(storageId) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid storage";
+                return RedirectToAction("Error", "Home");
+            }
+
+            string[] accessoriesCategory = Enum.GetNames(typeof(CategoryAccessories));
+
+            if (!accessoriesCategory.Contains(category))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid category";
+                return RedirectToAction("Error", "Home");
+            }
+
+            var storage = await storageService.GetStorageById(storageId);
+            int familyId = await familyService.GetFamilyId(User.Id());
+
+            if (storage.House.FamilyId != familyId)
+            {
+                TempData[MessageConstant.WarningMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
             var model = new AddAccessoriesViewModel()
             {
                 StorageId = storageId,
@@ -57,6 +108,28 @@ namespace WardrobeOrganizer.Controllers
                 TempData[MessageConstant.ErrorMessage] = "Something went wrong! Try again";
                 return View(model);
             }
+            if (await storageService.ExistsById(model.StorageId) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid storage";
+                return RedirectToAction("Error", "Home");
+            }
+            string[] accessoriesCategory = Enum.GetNames(typeof(CategoryAccessories));
+
+            if (!accessoriesCategory.Contains(model.Category))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid category";
+                return RedirectToAction("Error", "Home");
+            }
+
+
+            var storage = await storageService.GetStorageById(model.StorageId);
+            int familiId = await familyService.GetFamilyId(User.Id());
+
+            if (storage.House.FamilyId != familiId)
+            {
+                TempData[MessageConstant.WarningMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
 
             var storageId = model.StorageId;
 
@@ -67,8 +140,7 @@ namespace WardrobeOrganizer.Controllers
             catch (Exception e)
             {
 
-                this.ModelState.AddModelError(string.Empty, e.Message);
-                return this.View(model);
+                TempData[MessageConstant.ErrorMessage] = "Something went wrong! Try again";
             }
 
             return RedirectToAction("AccessoriesByCategory", "Accessories", new { storageId, model.Category });
@@ -78,6 +150,28 @@ namespace WardrobeOrganizer.Controllers
         [HttpGet]
         public async Task<IActionResult> AccessoriesByCategory(int storageId, string category)
         {
+            if (await storageService.ExistsById(storageId) == false)
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid storage";
+                return RedirectToAction("Error", "Home");
+            }
+            string[] accessoriesCategory = Enum.GetNames(typeof(CategoryAccessories));
+
+            if (!accessoriesCategory.Contains(category))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid category";
+                return RedirectToAction("Error", "Home");
+            }
+
+
+            var storage = await storageService.GetStorageById(storageId);
+            int familiId = await familyService.GetFamilyId(User.Id());
+
+            if (storage.House.FamilyId != familiId)
+            {
+                TempData[MessageConstant.WarningMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
             try
             {
                 var model = await accessoriesService.AllAccessoriesByCategory(storageId, category);
@@ -92,6 +186,23 @@ namespace WardrobeOrganizer.Controllers
 
         public async Task<IActionResult> Details(int accessoriesId)
         {
+            if (await accessoriesService.ExistsById(accessoriesId) == false)
+            {
+                TempData[MessageConstant.WarningMessage] = "Can`t find this shoes";
+                return RedirectToAction("Error", "Home");
+            }
+
+            var accessories = await accessoriesService.GetAccessoriesDetailsModelById(accessoriesId);
+            var house = await houseService.GetHouseById(accessories.HouseId);
+
+            int familiId = await familyService.GetFamilyId(User.Id());
+            var user = await userManager.FindByIdAsync(User.Id());
+
+            if (house.FamilyId != familiId && await userManager.IsInRoleAsync(user, RoleConstants.User))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
             try
             {
                 var model = await accessoriesService.GetAccessoriesDetailsModelById(accessoriesId);
@@ -113,14 +224,22 @@ namespace WardrobeOrganizer.Controllers
                 ModelState.AddModelError("", "Accessorie does not exist");
                 return RedirectToAction("Home", "Error");
             }
-
             var accessorie = await accessoriesService.GetAccessoriesDetailsModelById(accessoriesId);
+            var house = await houseService.GetHouseById(accessorie.HouseId);
 
+            int familyId = await familyService.GetFamilyId(User.Id());
+            var user = await userManager.FindByIdAsync(User.Id());
+
+            if (house.FamilyId != familyId && await userManager.IsInRoleAsync(user, RoleConstants.User))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
+          
             try
             {
                 await accessoriesService.DeleteById(accessoriesId);
-                TempData[MessageConstant.ErrorMessage] = "Accessorie was deleted";
-                
+                TempData[MessageConstant.ErrorMessage] = "Accessorie was deleted";         
             }
             catch (Exception)
             {
@@ -140,9 +259,18 @@ namespace WardrobeOrganizer.Controllers
                 ModelState.AddModelError("", "Accessorie does not exist");
                 return RedirectToAction("Home", "Error");
             }
-            var accessorise = await accessoriesService.GetAccessoriesEditModelById(accessoriesId);
+            var accessorise = await accessoriesService.GetAccessoriesDetailsModelById(accessoriesId);
+            var house = await houseService.GetHouseById(accessorise.HouseId);
 
-            var familyId = await familyService.GetFamilyId(User.Id());
+            int familyId = await familyService.GetFamilyId(User.Id());
+            var user = await userManager.FindByIdAsync(User.Id());
+
+            if (house.FamilyId != familyId && await userManager.IsInRoleAsync(user, RoleConstants.User))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
+
             var model = new EditAccessoriesViewModel()
             {
                 Id = accessoriesId,
@@ -171,7 +299,17 @@ namespace WardrobeOrganizer.Controllers
                 TempData[MessageConstant.ErrorMessage] = "Something went wrong! Try again";
                 return View(model);
             }
+            var accessorise = await accessoriesService.GetAccessoriesDetailsModelById(model.Id);
+            var house = await houseService.GetHouseById(accessorise.HouseId);
 
+            int familyId = await familyService.GetFamilyId(User.Id());
+            var user = await userManager.FindByIdAsync(User.Id());
+
+            if (house.FamilyId != familyId && await userManager.IsInRoleAsync(user, RoleConstants.User))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
 
             try
             {
@@ -189,6 +327,19 @@ namespace WardrobeOrganizer.Controllers
 
         public async Task<IActionResult> MemberAllAccessories(int memberId)
         {
+            if (await memberService.ExistsById(memberId))
+            {
+                return RedirectToAction("Error", "Home");
+            }
+
+            var member = await memberService.GetMemberById(memberId);
+            var familyId = await familyService.GetFamilyId(User.Id());
+
+            if (member.Family.Id != familyId)
+            {
+                return RedirectToAction("Error", "Home");
+
+            }
             try
             {
                 var model = await accessoriesService.AllAccessoriesByMemberId(memberId);
@@ -204,6 +355,27 @@ namespace WardrobeOrganizer.Controllers
 
         public async Task<IActionResult> MemberAccessoriesByCategory(int memberId, string category)
         {
+            if (await memberService.ExistsById(memberId))
+            {
+                return RedirectToAction("Error", "Home");
+            }
+
+            string[] accessoriesCategory = Enum.GetNames(typeof(CategoryAccessories));
+
+            if (!accessoriesCategory.Contains(category))
+            {
+                TempData[MessageConstant.ErrorMessage] = "Not valid category";
+                return RedirectToAction("Error", "Home");
+            }
+
+            var member = await memberService.GetMemberById(memberId);
+            int familiId = await familyService.GetFamilyId(User.Id());
+
+            if (member.Id != familiId)
+            {
+                TempData[MessageConstant.WarningMessage] = "Not allowed";
+                return RedirectToAction("Error", "Home");
+            }
             try
             {
                 var model = await accessoriesService.AllAccessoriesByCategoryAndMemberId(memberId, category);
